requirements.update.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
allennlp==2.10.1
annotated-types==0.7.0
anyio==3.4.0
async-timeout==4.0.3
attrs==24.2.0
base58==2.1.1
blinker==1.8.2
blis==1.0.1
boto3==1.35.37
botocore==1.35.37
cached-path==1.1.6
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
commonmark==0.9.1
confection==0.1.5
cymem==2.0.8
dill==0.3.9
discord.py==2.0.1
distro==1.9.0
docker-pycreds==0.4.0
exceptiongroup==1.2.2
fairscale==0.4.6
filelock==3.7.1
Flask==2.2.5
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.21.0
google-auth==2.35.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
greenlet==3.1.1
h11==0.14.0
h5py==3.12.1
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.10.1
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.6.1
jmespath==1.0.1
joblib==1.4.2
jsonnet==0.20.0
langcodes==3.4.1
language_data==1.2.0
lmdb==1.5.1
marisa-trie==1.2.0
MarkupSafe==3.0.1
more-itertools==10.5.0
multidict==6.1.0
murmurhash==1.0.10
nltk==3.9.1
numpy==2.0.2
openai==1.51.2
packaging==24.1
pathlib_abc==0.1.1
pathtools==0.1.2
pathy==0.11.0
pillow==10.4.0
pluggy==1.5.0
preshed==3.0.9
promise==2.3
propcache==0.2.0
proto-plus==1.24.0
protobuf==3.20.3
psutil==6.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==1.9.0
pydantic_core==2.23.4
Pygments==2.18.0
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.1.1
regex==2024.9.11
requests==2.32.3
rich==12.6.0
rsa==4.9
s3transfer==0.10.3
sacremoses==0.1.1
scikit-learn==1.5.2
scipy==1.13.1
sentencepiece==0.2.0
sentry-sdk==2.16.0
setproctitle==1.3.3
shortuuid==1.0.13
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.1
spacey==0.1.1
spacy==3.3.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spotipy==2.24.0
SQLAlchemy==2.0.35
srsly==2.4.8
tensorboardX==2.6.2.2
termcolor==1.1.0
thinc==9.1.1
threadpoolctl==3.5.0
tokenizers==0.12.1
tomli==2.0.2
torch==1.12.1
torchvision==0.13.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.20.1
typer==0.4.2
typing_extensions==4.12.2
urllib3==1.26.20
wandb==0.12.21
wasabi==0.10.1
Werkzeug==3.0.4
yarl==1.14.0
zipp==3.20.2
